+++ modules/mystic-ai/entry.cpp
@@ -55,7 +55,7 @@ extern "C" __declspec(dllexport) AddonDefinition_t* GetAddonDef()
 
     AddonDef.Name             = "Mystic AI";
     AddonDef.Version.Major    = 1;
-    AddonDef.Version.Minor    = 0;
+    AddonDef.Version.Minor    = 1;
     AddonDef.Version.Build    = 0;
     AddonDef.Version.Revision = 0;
     AddonDef.Author      = "OgMorrow2090";